conformance/basicoperations/cl_commandQueue_enqueueNDRangeKernel.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following conditions
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLCommandQueue enqueueNDRangeKernel conformance test.</title>
    <link rel="stylesheet" href="../../resources/js-test-style.css"/>
    <script src="../../resources/js-test-pre.js"></script>
    <script src="../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script id="square" type="x-kernel">
__kernel void square(
    __global float* input,
    __global float* output,
    const unsigned int count)
{
    int i = get_global_id(0);
    if (i < count)
        output[i] = input[i] * input[i];
}
</script>
</script>
<script>
try {
    debug("WebCLCommandQueue enqueueNDRangeKernel test case.");
    debug("(WebCLKernel kernel, CLuint[3]? globalWorkOffset, CLuint[3]? globalWorkSize, CLuint[3]? localWorkSize, optional WebCLEvent[]? eventWaitList = null, optional WebCLEvent? event = null)\n");

    var SIZE = 1024;
    var count = SIZE;
    var BUFFER_SIZE = Float32Array.BYTES_PER_ELEMENT * 320 * 240 * 4;
    var data = new Float32Array(SIZE);
    for (var i = 0; i < count; i++) {
        data[i] = Math.random();
    }

    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var kernelSource = wtu.getKernel("square");
    var webCLProgram = wtu.createProgram(webCLContext, kernelSource);
    var webPlatforms = wtu.getPlatforms();
    var webCLDevices = wtu.getDevices(webCLPlatforms[0], webcl.DEVICE_TYPE_GPU);
    wtu.build(webCLProgram, webCLDevices, "-cl-single-precision-constant");
    var webCLKernel = wtu.createKernel(webCLProgram, "square");
    var webCLCommandQueue = wtu.createCommandQueue(webCLContext);
    var webCLEvent = wtu.createEvent(webCLContext);

    var localWorkSize = new Int32Array([workGroupSize]);
    var globalWorkSize = new Int32Array([count]);

    input = webCLContext.createBuffer(webcl.MEM_READ_ONLY, BUFFER_SIZE);
    output = webCLContext.createBuffer(webcl.MEM_WRITE_ONLY, BUFFER_SIZE);
    webCLCommandQueue.enqueueWriteBuffer(input, true, 0, DATA_SIZE, data);

    wtu.setArg(webCLKernel, 0, input);
    wtu.setArg(webCLKernel, 1, output);
    wtu.setArg(webCLKernel, 2, count, WebCLKernelArgumentTypes.LONG);

    var workGroupSize = webCLKernel.getWorkGroupInfo(webCLDevices[0], webcl.KERNEL_WORK_GROUP_SIZE);

    // FIXME :: Need to add test cases passing event parameters.

    var globalWorkSize = new Int32Array([300, 400]);
    var localWorkSize = new Int32Array([2, 2]);
    var globalWorkOffset = new Int32Array([2, 2]);

    shouldBeTrue("wtu.enqueueNDRangeKernel(webCLCommandQueue, webCLKernel, globalWorkOffset, globalWorkSize, localWorkSize);");
    shouldBeTrue("wtu.enqueueNDRangeKernel(webCLCommandQueue, webCLKernel, null, globalWorkSize, localWorkSize);");
    shouldBeTrue("wtu.enqueueNDRangeKernel(webCLCommandQueue, webCLKernel, globalWorkOffset, globalWorkSize, null);");
    shouldBeTrue("wtu.enqueueNDRangeKernel(webCLCommandQueue, webCLKernel, null, globalWorkSize, null);");

    // Tesitng for Exceptions.
    shouldThrowExceptionName("wtu.enqueueNDRangeKernel(webCLCommandQueue, invalid_object, globalWorkOffset, globalWorkSize, localWorkSize);", "INVALID_KERNEL");
    shouldThrowExceptionName("wtu.enqueueNDRangeKernel(webCLCommandQueue, webCLKernel, invalid_object, globalWorkSize, localWorkSize);", "INVALID_VALUE");
    shouldThrowExceptionName("wtu.enqueueNDRangeKernel(webCLCommandQueue, webCLKernel, globalWorkOffset, invalid_object, localWorkSize);", "INVALID_WORK_GROUP_SIZE");
    shouldThrowExceptionName("wtu.enqueueNDRangeKernel(webCLCommandQueue, webCLKernel, globalWorkOffset, globalWorkSize, invalid_object);", "INVALID_WORK_GROUP_SIZE");

} catch(e) {
    testFailed("WebCLCommandQueue :: enqueueNDRangeKernel() throws an exception : " + e.message);
}
notifyFinishedToHarness();
successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
